// apps/api/src/types/api.ts
export enum CurrencyCode {
  XOF = 'XOF',
  USD = 'USD',
  EUR = 'EUR',
  GHS = 'GHS',
  NGN = 'NGN',
  KES = 'KES',
  MRO = 'MRO',
}
// ...
export enum BillingFrequency {
  weekly = 'weekly',
  bi_weekly = 'bi-weekly',
  monthly = 'monthly',
  bi_monthly = 'bi-monthly',
  quarterly = 'quarterly',
  semi_annual = 'semi-annual',
  yearly = 'yearly',
}

export enum Frequency {
  weekly = 'weekly',
  bi_weekly = 'bi-weekly',
  monthly = 'monthly',
  bi_monthly = 'bi-monthly',
  quarterly = 'quarterly',
  semi_annual = 'semi-annual',
  yearly = 'yearly',
}

export enum FailedPaymentAction {
  cancel = 'cancel',
  pause = 'pause',
  continue = 'continue',
}

export enum FirstPaymentType {
  initial = 'initial',
  non_initial = 'non_initial',
}
// ...
export interface SubscriptionPlan extends Record<string, unknown> {
  // Extends CreateSubscriptionPlan
  plan_id?: string;
  merchant_id?: string;
  organization_id?: string;
  created_at?: Date;
  updated_at?: Date;
  [key: string]: unknown;
}